feat: refactor jvm_instance.h for gdextension

+++ src/api/script/jvm_placeholder_instance.cpp
@@ -22,7 +22,7 @@ Variant JvmPlaceHolderInstance::callp(const StringName& p_method, const Variant*
             packed.append("This script can't be found in your JVM project. Don't forget to build it and use a valid "
                           "gdj/kt/java file.");
             return packed;
-        } else if (script_ref->get_last_time_source_modified() > static_cast<uint64_t>(JvmScriptManager::get_instance()->get_last_reload())) {
+        } else if (script_ref->get_last_time_source_modified() > JvmScriptManager::get_instance()->get_last_reload()) {
             PackedStringArray packed {};
             packed.append("This script has been modified since the last time you built your project.");
             return packed;
